import bug fixes
