+++ lstore/query.py
@@ -1,4 +1,4 @@
-from typing import Literal
+from typing import Literal,Union
 
 from lstore.table import Table, Record
 from lstore.index import Index
@@ -55,7 +55,7 @@ class Query:
         search_key,
         search_key_index,
         projected_columns_index: list[Literal[0, 1]]
-    ) -> list[Record] | False:
+    ) -> Union[list[Record], bool]:
         """
         # Read matching record with specified search key
         # :param search_key: the value you want to search based on
