+++ lstore/storage/record_index.py
@@ -1,0 +1,3 @@
+from collections import namedtuple
+
+RecordIndex = namedtuple("RecordIndex", ["page_id", "offset"])
